Split things up
